search

+++ changgou_parent/changgou_service/changgou_goods_service/src/main/java/com/tabwu/changgou/goods/service/AlbumService.java
@@ -11,7 +11,7 @@ public interface AlbumService {
 
     void updateAlbum(Album album);
 
-    Album findOne(Integer id);
+    Album findOne(Long id);
 
     PageResult findAll(Integer currentPage, Integer pageSize);
